Refactor: unify the implementation of OutsideWorkspaceMetadata and go_back_to_integration
