+++ crates/but-workspace/src/lib.rs
@@ -46,7 +46,7 @@ pub mod commit_engine;
 pub mod tree_manipulation;
 pub use tree_manipulation::function::discard_workspace_changes;
 pub mod head;
-pub use head::head;
+pub use head::{head, merge_worktree_with_workspace};
 
 /// 🚧utilities for applying and unapplying branches 🚧.
 pub mod branch;
